Use helper fab_exists instead of os.path.isfile.
The use of os.path.isfile will work if we are in the local environement but will
not work if we need to know if the file exist in a remote environement. fab_exists will
use the correct way to findout if the file exists using the role env variable.