WIP Refactoring to made drupalizer work with docker-compose.
I'm really sorry for this one BIG commit with all the changes, but this
start like an experiment work testing differents thinks and some of
them were removed, so I don't want to create confussion with commits
that add some changes and are later removed by other commits.
This is still a WIP (Work In Progress) but I think is ready for
testing and have some feedbacks and contributions!

* Use Docker4Drupal instead our own base image.
* Use Docker Compose to run docker commands instead docker directly, so
almost every fab docker.* task has changed, use fab --list to see new tasks.
* Add a fix permissions task to set the owner and groups of files
dynamically using the current user and the web server user in the container.
* Add a task to clean the src/drupal avoiding permissions problems.
* Add ruamel.yml like a dependency to the project to inspect the defined
services in the docker-compose.yml
* Change the test package to load and run differents tests modules dynamically.
* Add a "tags" argument to the "test" command in order to limit which
scenarios are run.
* Changes in the config files structure, you must copy local_vars.py and
rewrite the conf/drupalizer.py
* Several PEP8 corrections.
* Add aegir folder with the drush migrate-sites and remove-platform scripts
that should be placed in the aegir home in the hostmaster server.